requirements.txt
opencv-python
Pillow
pytest>=4.5.0
PyQt5